improve vectorization of minhash
