+++ Cargo.toml
@@ -113,32 +113,33 @@ tikv-jemallocator = {version = "0.5.4", features = [
 [workspace]
 members = [
   "src/arrow2",
-  "src/parquet2",
+  "src/common/daft-config",
   "src/common/display",
   "src/common/error",
   "src/common/io-config",
-  "src/common/treenode",
-  "src/common/daft-config",
   "src/common/system-info",
+  "src/common/treenode",
   "src/daft-core",
-  "src/daft-local-execution",
-  "src/daft-io",
-  "src/daft-image",
-  "src/daft-parquet",
   "src/daft-csv",
-  "src/daft-json",
   "src/daft-dsl",
-  "src/daft-table",
-  "src/daft-plan",
-  "src/daft-physical-plan",
+  "src/daft-functions",
+  "src/daft-functions-json",
+  "src/daft-hash",
+  "src/daft-image",
+  "src/daft-io",
+  "src/daft-json",
+  "src/daft-local-execution",
   "src/daft-micropartition",
+  "src/daft-parquet",
+  "src/daft-physical-plan",
+  "src/daft-plan",
   "src/daft-scan",
   "src/daft-scheduler",
   "src/daft-sketch",
-  "src/daft-functions",
-  "src/daft-functions-json",
   "src/daft-sql",
-  "src/hyperloglog"
+  "src/daft-table",
+  "src/hyperloglog",
+  "src/parquet2"
 ]
 
 [workspace.dependencies]
@@ -155,6 +156,7 @@ bytes = "1.6.0"
 chrono = "0.4.38"
 chrono-tz = "0.8.4"
 comfy-table = "7.1.1"
+daft-hash = {path = "src/daft-hash"}
 derivative = "2.2.0"
 divan = "0.1.14"
 dyn-clone = "1"
@@ -166,6 +168,7 @@ jaq-core = "1.2.0"
 jaq-interpret = "1.2.0"
 jaq-parse = "1.0.0"
 jaq-std = "1.2.0"
+mur3 = "0.1.0"
 num-derive = "0.3.3"
 num-traits = "0.2"
 once_cell = "1.19.0"
@@ -177,6 +180,7 @@ regex = "1.10.4"
 rstest = "0.18.2"
 rustc-hash = "2.0.0"
 serde_json = "1.0.116"
+sha1 = "0.11.0-pre.4"
 sketches-ddsketch = {version = "0.2.2", features = ["use_serde"]}
 snafu = {version = "0.7.4", features = ["futures"]}
 sqlparser = "0.51.0"
