+++ daft/daft/__init__.pyi
@@ -1211,6 +1211,7 @@ def minhash(
     num_hashes: int,
     ngram_size: int,
     seed: int = 1,
+    hash_function: Literal["murmur3", "xxhash", "sha1"] = "murmur3",
 ) -> PyExpr: ...
 
 # -----
@@ -1346,7 +1347,13 @@ class PySeries:
     def sort(self, descending: bool) -> PySeries: ...
     def argsort(self, descending: bool) -> PySeries: ...
     def hash(self, seed: PySeries | None = None) -> PySeries: ...
-    def minhash(self, num_hashes: int, ngram_size: int, seed: int = 1) -> PySeries: ...
+    def minhash(
+        self,
+        num_hashes: int,
+        ngram_size: int,
+        seed: int = 1,
+        hash_function: Literal["murmur3", "xxhash", "sha1"] = "murmur3",
+    ) -> PySeries: ...
     def __invert__(self) -> PySeries: ...
     def count(self, mode: CountMode) -> PySeries: ...
     def sum(self) -> PySeries: ...
