+++ src/daft-functions/Cargo.toml
@@ -7,15 +7,19 @@ common-io-config = {path = "../common/io-config", default-features = false}
 common-runtime = {path = "../common/runtime", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
+daft-hash = {workspace = true}
 daft-image = {path = "../daft-image", default-features = false}
 daft-io = {path = "../daft-io", default-features = false}
 futures = {workspace = true}
+mur3 = {workspace = true}
 paste = "1.0.15"
 pyo3 = {workspace = true, optional = true}
+sha1 = "0.11.0-pre.4"
 tiktoken-rs = {workspace = true}
 tokio = {workspace = true}
 typetag = "0.2.16"
 uuid = "1.10.0"
+xxhash-rust = {workspace = true, features = ["xxh64"]}
 bytes.workspace = true
 serde.workspace = true
 snafu.workspace = true
